+++ UOP2020-JANUAR/src/guiFromeZaDodavanjeiIzmene/ServiserForma.java
@@ -112,7 +112,7 @@ public class ServiserForma extends JFrame {
 					int telefon = Integer.parseInt(txtTelefon.getText().trim());
 					String korisnickoIme = txtKorisnickoIme.getText().trim();
 					String lozinka = new String(pfLozinka.getPassword()).trim();
-					Float plata = Float.parseFloat(txtPlata.getText().trim());
+					Double plata = Double.parseDouble(txtPlata.getText().trim());
 					Specijalizacija specijalizacija = (Specijalizacija)cbSpecijalizacija.getSelectedItem();
